pyproject.toml
[tool.poetry]
name = "smart_context_reducer"
version = "0.1.0"
description = ""
authors = ["cybrvybe <alexfigueroa.cybr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.10.5"
asyncio = "^3.4.3"
tqdm = "^4.66.5"
colorama = "^0.4.6"
python-dotenv = "^1.0.1"
pyperclip = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
copy-context = "copy_context:cli_main"
